Beginning path following
Now beginning to integrate the path-planning with the locomotion. mujoco_path_visualizer.py updates the base xml file with the waypoints for path planning. walk_to_multiple_locationsv2.py is the main file which gets the waypoints from mujoco_path_visualizer.py which are then fed to the quadruped controller. :)